Refactor includes
The main idea here is to simplify the includes
and to prepare the shift to new versions of
bindings. To achieve that goal and because
further compatibility with versions 1, 2 and vdyn
are not required, the predefs and postdefs are
removed, the verbose functions are made common
and the order of defines is reverted.

Signed-off-by: José Bollo <jose.bollo@iot.bzh>