Improve effect of defining AFB_BINDING_X4R1_ITF_REVISION
The effect of AFB_BINDING_X4R1_ITF_REVISION is to restrict
the exposed interface for compiling bindings for older
versions of the binder.

That mechanism was not complete because it did not forbid to
use functions of the unrestricted interface. This commit fixes
that issue by really removing part of the interface that are
not available for the selected revision.