Fix C++ tutorial example
The main development is running C. So the C++ tutorial
has remained unfinished since long time.

This commit fixes serious bugs that were making tutorial
not useable.

The state after that commit is better but there is still
work to be done for making that tutorial safe. For example,
the management of reference count of data is not clear and
should be implemented in C++ automatically within afb::data
class.